+++ Commands/lib/aphid/rake.rb
@@ -2,4 +2,5 @@
 
 require File.join(File.dirname(__FILE__), "extensions")
 require File.join(File.dirname(__FILE__), "rake", "support")
+require File.join(File.dirname(__FILE__), "rake", "publish")
 require File.join(File.dirname(__FILE__), "rake", "tasks")