+++ Skeleton/Publish.rb
@@ -1,0 +1,9 @@
+#
+# Aphid "Publish" Configuration
+#
+
+{
+  # :hosts => [ "web1.example.com", "web2.example.com" ],
+  # :path  => "/srv/application",
+  # :user  => "jdoe"
+}